requirements.txt
# SpendWise AI - Requirements
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.25.0
anthropic>=0.18.0
tqdm>=4.65.0
Pillow>=9.5.0
python-dotenv>=1.0.0
datasets>=2.14.0
pytesseract>=0.3.10
